refactor(server): extract CombatSystem from ArenaRoom
- Extract combat logic to server/src/rooms/systems/combatSystem.ts
- Functions moved: processCombat, applyShieldReflection, spawnPvPBiteOrbs,
  applyProjectileDamage, applySelfDamage, explodeBomb
- ArenaRoom.ts reduced by 331 lines (3749 → 3418)
- All tests pass, determinism preserved

Part 1 of ArenaRoom decomposition (slime-arena-ar1)
